fix: mypy (type-only, no behaviour change)
Clears the 14 type errors the typecheck CI flagged, all from the genomes merge:
- color: the colormap sampler returns interpolated floats, so annotate it
  tuple[float, float, float] and return an explicit 3-tuple; map_values can
  return a None scale, so its type is tuple[dict, dict | None].
- render.as_svg / the two _nice_step helpers / zombi._genomes_dir: pin the
  Any-typed returns (str(...), float(...), a typed local).
- trees.figure: a leaf name may be None -> TipPos gets `name or ""`.
- genomes.figure._clone: the **kw rebuild is checked at the call sites, so
  silence the untyped-dict arg-type with a scoped ignore.

+++ src/phylustrator/genomes/layers/guides.py
@@ -64,7 +64,7 @@ def _nice_step(span: float, target: int) -> float:
     exp = math.floor(math.log10(raw))
     base = raw / 10 ** exp
     nice = 1 if base < 1.5 else 2 if base < 3.5 else 5 if base < 7.5 else 10
-    return nice * 10 ** exp
+    return float(nice * 10 ** exp)
 
 
 def _fmt_bp(v: float) -> str:
